requirements.txt
astroid==2.11.4
attrs==21.4.0
bandit==1.7.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colour==0.1.5
coverage==6.3.2
dill==0.3.4
distlib==0.3.4
et-xmlfile==1.1.0
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.27
identify==2.5.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==4.5.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector-python==8.0.29
nodeenv==1.6.0
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pbr==5.9.0
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
protobuf==3.20.1
py==1.11.0
pylint==2.13.8
pyotp==2.6.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
six==1.16.0
smmap==5.0.0
stevedore==3.5.0
styleframe==4.0.0
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
tox==3.25.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wrapt==1.14.1
xlrd==1.2.0
